+++ integrationtests/as-integration-embedded/src/test/java/org/infinispan/test/integration/as/cdi/GreetingCacheManager.java
@@ -39,7 +39,7 @@ public class GreetingCacheManager {
       return cache.getCacheConfiguration().eviction().strategy();
    }
 
-   public int getEvictionMaxEntries() {
+   public long getEvictionMaxEntries() {
       return cache.getCacheConfiguration().eviction().maxEntries();
    }
